+++ src/main/scala/edu/ie3/simona/model/participant2/HpModel.scala
@@ -218,24 +218,8 @@ class HpModel private (
           )
       }
 
-    val (newHpActivePower, _, qDotIntoGrid) = {
-      if (turnOn)
-        (pRated, pThermal, pThermal)
-      else if (
-        currentStorageEnergy > zeroKWh && state.thermalDemands.houseDemand.hasRequiredDemand
-      ) {
-        // If the house has req. demand and storage isn't empty, we can heat the house from storage.
-        (zeroKW, zeroKW, storagePThermal)
-      } else if (
-        currentStorageEnergy > zeroKWh && state.thermalDemands.houseDemand.hasPossibleDemand && state.lastHpOperatingPoint.thermalOps.qDotHouse > zeroKW
-      )
-        // Edge case when em controlled: If the house was heated last state by Hp and setPower is below turnOn condition now,
-        // but house didn't reach target or boundary temperature yet. House can be heated from storage, if this one is not empty.
-        (zeroKW, zeroKW, storagePThermal)
-      else (zeroKW, zeroKW, zeroKW)
-    }
-
-    (newHpActivePower, qDotIntoGrid)
+    if (turnOn) (pRated, pThermal)
+    else (zeroKW, zeroKW)
   }
 
   override def createResults(
@@ -299,7 +283,6 @@ class HpModel private (
       if (qDotIntoGrid > zeroKW) {
         thermalGrid.handleFeedIn(
           state,
-          newActivePowerHp > zeroKW,
           qDotIntoGrid,
           state.thermalDemands,
         )
@@ -337,7 +320,6 @@ class HpModel private (
       if (qDotIntoGrid > zeroKW) {
         thermalGrid.handleFeedIn(
           state,
-          newActivePowerHp > zeroKW,
           qDotIntoGrid,
           state.thermalDemands,
         )